+++ Runtime/SpookySceneLoader.cs
@@ -1,15 +1,21 @@
-using UnityEditor.Events;
+using IAmTaku.SpookyLoader.Triggers;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 
-namespace Spooky_Loader
+namespace IAmTaku.SpookyLoader
 {
     public class SpookySceneLoader : MonoBehaviour
     {
+        #region FIELDS
+
         public bool isLoaded;
         [Tooltip("Add a prefab off your trigger to quickly reuse it.")]
-        public GameObject _triggerPrefab;
-        
+        public SpookyTrigger spookyTriggerPrefab;
+
+        #endregion
+
+        #region METHODS
+
         public void OnSceneTrigger()
         {
             if (!isLoaded)
@@ -19,34 +25,19 @@ namespace Spooky_Loader
             }
             UnLoadScene();
         }
-        
+
         private void LoadScene()
         {
             SceneManager.LoadSceneAsync(gameObject.name, LoadSceneMode.Additive);
             isLoaded = true;
         }
-        
+
         private void UnLoadScene()
         {
-            SceneManager.UnloadSceneAsync(gameObject.name); 
+            SceneManager.UnloadSceneAsync(gameObject.name);
             isLoaded = false;
         }
 
-#if UNITY_EDITOR
-        public void CreateTrigger()
-        {
-            var _sceneTrigger = Instantiate(_triggerPrefab, transform);
-            _sceneTrigger.name = "SpookyTrigger";
-            _sceneTrigger.transform.position = transform.position;
-            _sceneTrigger.transform.rotation = transform.rotation;
-            _sceneTrigger.transform.localScale = transform.localScale;
-            
-            UnityEventTools.AddPersistentListener(_sceneTrigger
-                .GetComponent<SpookyTrigger>()
-                .onSceneTrigger, OnSceneTrigger);
-            
-        } 
-#endif
-        
+        #endregion
     }
 }
