+++ Runtime/Triggers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9af0e0b76d51c7845ac18d589e967121
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
